% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bIndicator.R
\name{bIndicator}
\alias{bIndicator}
\title{Indicator for binomial RDT}
\usage{
bIndicator(pi, R)
}
\arguments{
\item{pi}{Failure probability.}

\item{R}{Lower Level reliability requirement.}
}
\value{
0 -- No; 1 -- Yes.
}
\description{
Determine whether the failure probability satisfies the lower level reliability requirement. (Only for binomial RDT)
}
\examples{
bIndicator(pi = 0.05, R = 0.9)
1
bIndicator(pi = 0.2, R = 0.9)
0

}
