% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bproducerrisk.R
\name{bproducerrisk}
\alias{bproducerrisk}
\title{Producer's Risk for Binomial RDT}
\usage{
bproducerrisk(n, c, pi, R)
}
\arguments{
\item{n}{RDT sample size.}

\item{c}{Maximum allowable failures.}

\item{pi}{Failure probability.}

\item{R}{Lower level reliability requirement.}
}
\value{
Probability for producer's risk
}
\description{
Define the producer's risk function which gets the probability of rejecting the test when the lower level reliability requirement can be satisfied (for binomial RDT).
}
\examples{
pi <- pi_MCSim_beta(M = 5000, seed = 10, a = 1, b = 1)
bproducerrisk(n = 10, c = 2, pi = pi, R = 0.8);
0.02826733
}
