% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brejectprob.R
\name{brejectprob}
\alias{brejectprob}
\title{Rejection Probability for Binomial RDT}
\usage{
brejectprob(n, c, pi)
}
\arguments{
\item{n}{RDT sample size.}

\item{c}{Maximum allowable failures.}

\item{pi}{Failure probability.}
}
\value{
Rejection probability
}
\description{
Define the rejection probability function which gets the probability of rejecting the test (for binomial RDT).
}
\examples{
pi <- pi_MCSim_beta(M = 5000, seed = 10, a = 1, b = 1)
brejectprob(n = 10, c = 2, pi = pi);
}
\seealso{
\code{\link{bconsumerrisk}}, \code{\link{bproducerrisk}}, \code{\link{bacceptprob}}
}
