% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_MCSim_beta.R
\name{pi_MCSim_beta}
\alias{pi_MCSim_beta}
\title{Define the simulation function to generate failure probability (for binomial)}
\usage{
pi_MCSim_beta(M, seed, a, b)
}
\arguments{
\item{M}{Simulation sample size}

\item{seed}{Random seed for random sample}

\item{a}{Shape parameter 1 for beta distribution}

\item{b}{Shape parameter 2 for beta distribution}
}
\value{
Vector of failure probability sample
}
\description{
Define the simulation function to generate failure probability (for binomial)
}
\examples{
pi_MCSim_beta(5000,10,1,1)
0.4925217968 0.5730923335 0.9148640311 ...
}
