% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_MCSim_dirichlet.R
\name{pi_MCSim_dirichlet}
\alias{pi_MCSim_dirichlet}
\title{Define the simulation function to generate failure probability (for multibinomial with dirichlet prior)}
\usage{
pi_MCSim_dirichlet(M, seed, par)
}
\arguments{
\item{M}{Simulation sample size}

\item{seed}{Random seed for random sample}

\item{par}{Parameters for dirichlet distribution}
}
\value{
Vector of failure probability sample
}
\description{
Define the simulation function to generate failure probability (for multibinomial with dirichlet prior)
}
\examples{
pi_MCSim_dirichlet(1,10,c(1,1,1))
0.5130724 0.3780014 0.1089262
}
